+++ src/components/BooksList.js
@@ -2,19 +2,8 @@ import React from "react"
 import Book from "./Book"
 import { Button } from "react-bootstrap"
 import "bootstrap/dist/css/bootstrap.min.css"
+import { bookObj } from "../pages"
 
-export const bookObj = [
-  {
-    id: 1,
-    title: "winds of winter",
-    category: "fantasy",
-  },
-  {
-    id: 2,
-    title: "game of thrones",
-    category: "horror",
-  },
-]
 const BooksList = () => {
   const displayBooks = bookObj.map(book => {
     return <Book key={book.id} book={book} />
