+++ src/pages/index.js
@@ -3,8 +3,18 @@ import { Link } from "gatsby"
 import App from "../components/App.js"
 import { Provider } from "react-redux"
 import store from "../reducers/books"
-import { bookObj } from "../components/BooksList"
-
+export const bookObj = [
+  {
+    id: Math.floor(Math.random() * 100),
+    title: "winds of winter",
+    category: "fantasy",
+  },
+  {
+    id: Math.floor(Math.random() * 100),
+    title: "game of thrones",
+    category: "horror",
+  },
+]
 const IndexPage = () => {
   const [books, setBooks] = useState(bookObj)
   return (
