+++ src/reducers/index.js
@@ -1,8 +1,6 @@
 import { combineReducers } from "redux"
 
 const reducer = (book = {}, action) => {
-  console.log(action)
-
   switch (action.type) {
     case "CREATE_BOOK":
       return { ...book, ...action.payload }
